+++ lib/utils/app_constants.dart
@@ -5,6 +5,14 @@ class AppConstants {
       "https://github.com/karnauhov/Revelation/blob/main/PrivacyPolicy.md";
   static const String licenseUrl =
       "https://github.com/karnauhov/Revelation/blob/main/LICENSE";
+  static const String projectUrl = "https://github.com/karnauhov/Revelation";
+  static const String latestReleaseUrl =
+      "https://github.com/karnauhov/Revelation/releases/latest";
+  static const String websiteUrl = "https://www.revelation.website";
+  static const String googlePlayUrl =
+      "https://play.google.com/store/apps/details?id=ai11.link.revelation";
+  static const String microsoftStoreUrl =
+      "https://apps.microsoft.com/detail/9NXHRR2P4087";
   static const languages = {
     'en': 'English',
     'uk': 'Українська',